refactor: components, constants, context 폴더구조 재구성
