+++ src/components/layout/GlobalErrorBoundary/GlobalErrorBoundary.tsx
@@ -1,7 +1,7 @@
 import { ErrorBoundary } from 'react-error-boundary';
-import { ErrorFallback } from '../shared/ErrorFallback';
+import { ErrorFallback } from '../../shared/ErrorFallback';
 import { useQueryErrorResetBoundary } from '@tanstack/react-query';
-import { HTTPError } from '../../apis/HTTPError';
+import { HTTPError } from '../../../apis/HTTPError';
 
 interface Props {
 	children: React.ReactNode;
