+++ src/components/layout/Header/components/NavLinks.tsx
@@ -1,9 +1,12 @@
 import styled from '@emotion/styled';
 import { Link } from 'react-router-dom';
-import { NAV_LINKS, CATEGORY_LINKS } from '../constants/navLinks';
+import {
+	NAV_LINKS,
+	CATEGORY_LINKS,
+} from '../../../../constants/navigation/navLinks';
 import Dropdown from './Dropdown';
-import { mediaQuery } from '../utils/mediaQuery';
-import { useDropdown } from '../hooks/useDropdown';
+import { mediaQuery } from '@/utils/mediaQuery';
+import { useDropdown } from '@/hooks/useDropdown';
 import { useRef } from 'react';
 import { useNavigation } from '@/hooks/useNavigation';
