+++ src/constants/domain/emailDomain.ts
