+++ src/constants/product/search.ts
